data/datacatalogue/scripts/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.5
idna==3.2
numpy==1.22.2
pandas==1.4.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.6